updated Dependency_XVM_Main_Files to 7.8.0_nightly_8765
@willster:
yes! sadly it´s a nightly.
but i need something that will work and not break the game when updating xvm configs and don´t want to deal with a usermod or branch...
either we need to completeley disable xvm when releasing database to public or hopefully the stable will be out.
plz inform me if i should disable it, when u want to release!