disabled battlechatfilter (scriptloader)
updating this to json will take too much time that we could spend on other urgent things. i would say we take a look at this after v2 is out.
in my opinion: even with the "facebook filterlist" users find a way to insult each other. and this means the file would grow bigger and bigger....