en DependencyGlobal_RelHax_LoadingSplash / rem *_ResMods
i´ve made a mistake before as i removed the old package to not have two same ones (one was disabled).
what i did caused an error while loading the db with the installer. i remember having this issue a while ago....
however....sry for that! but after reverting that and removing the newer package it all is working again.