upd contour/tank-icons maintained by relhax (missing ones)
this commit finishes the update of all tank icons the relhax-team is maintaining.

also we have a TIM-repo that should get used: it doesnt creates unused icons (eg for tutorial).
to keep all releases built the same way, it would be better for the whole team to use the same program (sadly automation can not be used).
for the special icons (we only have one: contouricons_wizard) it would be no problem to add the special configurated TIM to the repo (preconfigured to just run the wizrad icons).
that would simplyfy sourcecontrol and every teammember has access to every tool relhax is using.
...just sayin´...