autoUpdate [xvm v8.9.0_0016]
@dirty20067
to update xvm, just run the "automation runner" (command line for shortcut: ...YOURDIRECTORY\RelhaxModpack.exe /automation-runner) and select the xvm packages from the list. then press start and the update process will beginn.
you can also do this for the xvm stable package
(because it is using the nightly build as default source -> in cases it should use the stable package you need to edit the root.xml. but beware: most of the time its better to keep the nightly source! so it better to keep it and only press the start button  ;) )